refactor(frontend): 🎨 socket architecture
better seperation of concerns between connection, event management and usage
