--- app/frontend/src/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from "./use-intersection-observer";
-export * from "./use-socket";
-export * from "./use-theme";
