+++ app/frontend/src/hooks/socket/use-ws.tsx
